backend/requirements.txt
fastapi
uvicorn
joblib
numpy
pandas
google-cloud-storage
google-auth
